few change
OPM for YMFM

+++ SDL/SDL3/ymfm_wrap.cpp
@@ -363,7 +363,7 @@ void OPM_Reset(void)
 }
 
 
-uint8_t FASTCALL OPM_Read(uint16_t adr)
+uint8_t FASTCALL OPM_Read(uint32_t adr)
 {
 	uint8_t ret = 0;
 	(void)adr;
